Reorganize where stuff is in the API in a way that makes a little more sense.
