+++ src/client/signals/options/enums.ts
